+++ Models/Print.cs
@@ -8,7 +8,7 @@ namespace Kostenberekening_3D_printer_ASPnet.Models
     public class Print : Filament
     {
         public decimal Duurtijd { get; set; }
-        public int AantalMeterPrint { get; set; }
+        public decimal AantalMeterPrint { get; set; }
         public decimal KostPerPrint { get; set; }
     }
 }
